+++ src/locale/lang/en.js
@@ -4,7 +4,7 @@ export default {
       now: 'Now',
       today: 'Today',
       cancel: 'Cancel',
-      // clear: 'Clear',
+      clear: 'Clear',
       confirm: 'OK',
       selectDate: 'Select date',
       selectTime: 'Select time',